Removed some compiler warnings.

+++ sw_stm32/FATFS/App/fatfs.cpp
@@ -47,7 +47,7 @@ void MX_FATFS_Init(void)
 DWORD get_fattime(void)
 {
   /* USER CODE BEGIN get_fattime */
-  return GNSS.get_FAT_time();
+  return (DWORD)(GNSS.get_FAT_time());
   /* USER CODE END get_fattime */
 }
